# docs/openapi.yaml
openapi: 3.0.1
info:
  title: ProductReview API
  description: API for create, read/retrieve, update and delete users, categories, reviews
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local server  
paths:
  /user/login:
    post:
      summary: User login
      description: Validate user credentials and return a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: User's username
                password:
                  type: string
                  description: User's password
              required:
                - email
                - password
      responses:
        200:
          description: User successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
        400:
          description: Invalid user credentials
          content:
            application/json:
              schema:
                type: string
              example: Bad Request
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /user/name:
    get:
      summary: Get user name by user ID
      description: Retrieve the user's name based on their ID. Requires a valid JWT token in the Authorization header.
      parameters:
        - name: id
          in: query
          description: The UUID of the user
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Bearer token for authorization
          required: true
          schema:
            type: string
      responses:
        200:
          description: User name retrieved successfully
          content:
            application/json:
              schema:
                type: string
                example: Jonas S.
        400:
          description: Invalid Authorization header or bad UUID format
          content:
            application/json:
              schema:
                type: string
              examples:
                missing_header:
                  value: Authorization header must be provided and start with 'Bearer '.
                bad_uuid:
                  value: ""
        401:
          description: Invalid or expired JWT token
          content:
            application/json:
              schema:
                type: string
              example: Invalid or expired JWT token.
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /categories/all:
    get:
      summary: Get all categories
      description: Retrieve a list of all category names. Requires a valid JWT token in the Authorization header.
      parameters:
        - name: Authorization
          in: header
          description: Bearer token for authorization
          required: true
          schema:
            type: string
      responses:
        200:
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Category name
        400:
          description: Invalid Authorization header
          content:
            application/json:
              schema:
                type: string
              example: Authorization header must be provided and start with 'Bearer '.
        401:
          description: Invalid or expired JWT token
          content:
            application/json:
              schema:
                type: string
              example: Invalid or expired JWT token.
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: string            
  /reviews/category:
    get:
      summary: Get reviews by category
      description: Retrieve reviews based on the specified category. Requires a valid JWT token in the Authorization header.
      parameters:
        - name: category
          in: query
          description: The category of reviews to retrieve
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Bearer token for authorization
          required: true
          schema:
            type: string
      responses:
        200:
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: Review ID
                    userId:
                      type: string
                      format: uuid
                      description: User ID associated with the review
                    categoryId:
                      type: integer
                      description: Category ID
                    productName:
                      type: string
                      description: Product name reviewed
                    reviewText:
                      type: string
                      description: Review text
                    rating:
                      type: integer
                      description: Rating given to the product
                    photo:
                      type: string
                      format: byte
                      description: Photo associated with the review
                    createdAt:
                      type: string
                      format: date-time
                      description: Timestamp when the review was created
        400:
          description: Invalid Authorization header
          content:
            application/json:
              schema:
                type: string
              example: Authorization header must be provided and start with 'Bearer '.
        401:
          description: Invalid or expired JWT token
          content:
            application/json:
              schema:
                type: string
              example: Invalid or expired JWT token.
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /reviews/add:
    post:
      summary: Add a new review
      description: Adds a review with an optional image
      parameters:
        - name: category
          in: query
          description: The review product category
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Bearer token for authentication
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                review:
                  type: string
                  description: JSON string containing review details (productName, reviewText, rating)
                  example: '{"productName": "Sample Product", "reviewText": "Great product!", "rating": 5}'
                image:
                  type: string
                  format: binary
                  description: Image file associated with the review
      responses:
        '200':
          description: Review added successfully
          content:
            application/json:
              schema:
                type: string
                example: success
        '400':
          description: Invalid authorization header
          content:
            application/json:
              schema:
                type: string
                example: Authorization header must be provided and start with 'Bearer '.
        '401':
          description: Invalid or expired JWT token
          content:
            application/json:
              schema:
                type: string
                example: Invalid or expired JWT token.
        '406':
          description: Review creation failed
          content:
            application/json:
              schema:
                type: string
                example: failed
                